Refactor database repository code. Create SHSRDAO interface for database collection. Create SubCollectionSHSRDAO for database sub collection.

+++ SHSR/src/main/java/com/SmartHealthRemoteSystem/SHSR/ReadSensorData/SensorDataRepository.java
@@ -1,9 +1,14 @@
 package com.SmartHealthRemoteSystem.SHSR.ReadSensorData;
 
+import com.SmartHealthRemoteSystem.SHSR.Repository.SHSRDAO;
+import com.SmartHealthRemoteSystem.SHSR.Repository.SubCollectionSHSRDAO;
+import com.SmartHealthRemoteSystem.SHSR.User.Doctor.Doctor;
 import com.SmartHealthRemoteSystem.SHSR.User.Patient.Patient;
 import com.SmartHealthRemoteSystem.SHSR.User.Patient.PatientRepository;
 import com.SmartHealthRemoteSystem.SHSR.User.Patient.PatientService;
+import com.SmartHealthRemoteSystem.SHSR.User.User;
 import com.google.api.core.ApiFuture;
+import com.google.cloud.Timestamp;
 import com.google.cloud.firestore.DocumentReference;
 import com.google.cloud.firestore.DocumentSnapshot;
 import com.google.cloud.firestore.Firestore;
@@ -13,25 +18,52 @@ import com.google.firebase.database.*;
 import lombok.SneakyThrows;
 import org.springframework.stereotype.Repository;
 
+import java.util.ArrayList;
+import java.util.Iterator;
+import java.util.List;
 import java.util.concurrent.ExecutionException;
 
 @Repository
-public class SensorDataRepository {
-    public static final String COL_NAME = "SensorData";
-    public final PatientRepository patientRepository;
+public class SensorDataRepository implements SHSRDAO<SensorData> {
+    private final String COL_NAME = "SensorData";
 
-    public SensorDataRepository(PatientRepository patientRepository) {
-        this.patientRepository = patientRepository;
-
-
+    @Override
+    public SensorData get(String sensorDataId) throws ExecutionException, InterruptedException {
+        Firestore dbFirestore = FirestoreClient.getFirestore();
+        DocumentReference documentReference = dbFirestore.collection(COL_NAME).document(sensorDataId);
+        ApiFuture<DocumentSnapshot> future = documentReference.get();
+        DocumentSnapshot document = future.get();
+        SensorData tempSensorData;
+        if (document.exists()) {
+            tempSensorData = document.toObject(SensorData.class);
+            return tempSensorData;
+        } else {
+            return null;
+        }
     }
 
-
-    public String CreateSensorData(SensorData sensorData)
-            throws InterruptedException, ExecutionException {
+    @Override
+    public List<SensorData> getAll() throws ExecutionException, InterruptedException {
         Firestore dbFirestore = FirestoreClient.getFirestore();
+        Iterable<DocumentReference> documentReference = dbFirestore.collection(COL_NAME).listDocuments();
+        Iterator<DocumentReference> iterator = documentReference.iterator();
 
+        List<SensorData> sensorDataList = new ArrayList<>();
+        SensorData sensorData;
+        while (iterator.hasNext()) {
+            DocumentReference documentReference1=iterator.next();
+            ApiFuture<DocumentSnapshot> future = documentReference1.get();
+            DocumentSnapshot document = future.get();
+            sensorData = document.toObject(SensorData.class);
+            sensorDataList.add(sensorData);
+        }
 
+        return sensorDataList;
+    }
+
+    @Override
+    public String save(SensorData sensorData) throws ExecutionException, InterruptedException {
+        Firestore dbFirestore = FirestoreClient.getFirestore();
         //auto create data ID by firebase
         DocumentReference addedDocRef = dbFirestore.collection(COL_NAME).document();
         sensorData.setSensorDataId(addedDocRef.getId());
@@ -42,82 +74,32 @@ public class SensorDataRepository {
         return addedDocRef.getId();
     }
 
-    public String UpdateSensorData(SensorData sensorData)
-            throws InterruptedException, ExecutionException {
+    @Override
+    public String update(SensorData sensorData) throws ExecutionException, InterruptedException {
         Firestore dbFirestore = FirestoreClient.getFirestore();
-
-        //auto create data ID by firebase
-
-
         DocumentReference addedDocRef = dbFirestore.collection(COL_NAME).document(sensorData.getSensorDataId());
-        ApiFuture<WriteResult> collectionsApiFuture =
-                dbFirestore.collection(COL_NAME).document(sensorData.getSensorDataId()).set(sensorData);
-        ApiFuture<WriteResult> writeResult = addedDocRef.update("timestamp", collectionsApiFuture.get().getUpdateTime());
-        return collectionsApiFuture.get().getUpdateTime().toString();
-    }
-
-    public SensorData getSensorDataDetails(String sensorDataId)
-            throws InterruptedException, ExecutionException {
-        Firestore dbFirestore = FirestoreClient.getFirestore();
-        DocumentReference documentReference = dbFirestore.collection(COL_NAME).document(sensorDataId);
-        ApiFuture<DocumentSnapshot> future = documentReference.get();
-        DocumentSnapshot document = future.get();
-        SensorData tempSensorData = null;
-        if (document.exists()) {
-            tempSensorData = document.toObject(SensorData.class);
-            return tempSensorData;
-        } else {
-            return null;
+        ApiFuture<WriteResult> collectionsApiFuture = null;
+        if(!(sensorData.getEcgReading().isEmpty())){
+            collectionsApiFuture =addedDocRef.update("ecgReading", sensorData.getEcgReading());
+        } else if (!(sensorData.getOxygenReading().isEmpty())){
+            collectionsApiFuture =  addedDocRef.update("oxygenReading", sensorData.getOxygenReading());
+        } else if (sensorData.getBodyTemperature() != null){
+            collectionsApiFuture = addedDocRef.update("bodyTemperature", sensorData.getBodyTemperature());
         }
+        if (collectionsApiFuture != null) {
+            ApiFuture<WriteResult> writeResult = addedDocRef.update("timestamp", collectionsApiFuture.get().getUpdateTime());
+            return writeResult.get().getUpdateTime().toString();
+        }
+        return Timestamp.now().toString();
     }
 
-    public String deleteSensorData(String sensorDataId) {
+    @Override
+    public String delete(String sensorDataId) throws ExecutionException, InterruptedException {
         Firestore dbFirestore = FirestoreClient.getFirestore();
+        if(get(sensorDataId) == null){
+            return "The sensorData with Id " + sensorDataId + " is not exist.";
+        }
         ApiFuture<WriteResult> writeResult = dbFirestore.collection(COL_NAME).document(sensorDataId).delete();
         return "Document with Sensor Data Id " + sensorDataId + " has been deleted";
     }
-
-    public void CreateTest(String PatientID)
-    {   String URL= "test/kesh";
-        // real-time database
-        final FirebaseDatabase database = FirebaseDatabase.getInstance();
-        DatabaseReference ref = database.getReference().child(URL);
-
-        ref.addValueEventListener(new ValueEventListener() {
-            @SneakyThrows
-            @Override
-            public void onDataChange(DataSnapshot dataSnapshot) {
-                String sensorData = dataSnapshot.getValue().toString();
-                SensorData test = new SensorData();
-                test.setEcgReading(sensorData);
-                String SensorDataID=CreateSensorData(test);
-                Patient ahmad = patientRepository.getPatient(PatientID);
-
-
-
-
-                if(ahmad.getSensorDataId().equals(""))
-                {
-                    ahmad.setSensorDataId(SensorDataID);
-                    patientRepository.updatePatient(ahmad);
-                }
-                else
-                {
-                    String Sen_ID= ahmad.getSensorDataId();
-                    test.setSensorDataId(Sen_ID);
-                    UpdateSensorData(test);
-                }
-
-
-            }
-            @Override
-            public void onCancelled(DatabaseError databaseError) {
-                System.out.println("The read failed: " + databaseError.getCode());
-
-            }
-        });
-    }
-
-
-
 }
